chore: make useModule2 a useModule
