+++ src/hooks/computer_vision/useClassification.ts
@@ -1,9 +1,9 @@
 import { ClassificationModule } from '../../modules/computer_vision/ClassificationModule';
 import { ResourceSource } from '../../types/common';
-import { useModule2 } from '../useModule2';
+import { useModule } from '../useModule';
 
 export const useClassification = ({
   modelSource,
 }: {
   modelSource: ResourceSource;
-}) => useModule2({ module: ClassificationModule, loadArgs: [modelSource] });
+}) => useModule({ module: ClassificationModule, loadArgs: [modelSource] });
